Refactor and clean up schedule code
